requirements.txt
redis>=5.0.1,<6.0.0
ollama>=0.3.3,<1.0.0
pyyaml>=6.0.1,<7.0.0
python-dotenv>=1.0.0,<2.0.0
numpy>=1.24.0,<2.0.0
# Vector embeddings and semantic search
sentence-transformers>=2.2.0,<3.0.0
redisvl>=0.0.1,<1.0.0
# Web interface
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
jinja2>=3.1.0,<4.0.0
python-multipart>=0.0.6,<1.0.0
# File handling and model management
aiofiles>=23.0.0,<24.0.0
pydantic>=2.0.0,<3.0.0
